Move data files to data/
