+++ data/fox.obj
